check approvedJob for not display in FindJob
display noDataImage fixed

+++ app/src/main/java/com/example/jobfinder/UI/FindNewJobs/FindNewJobViewModel.kt
@@ -29,37 +29,50 @@ class FindNewJobViewModel : ViewModel() {
     val jobsListLiveData: LiveData<List<JobModel>> get() = _jobsListLiveData
 //    val sortedJobsLiveData: LiveData<List<JobModel>> get() = _sortedJobsLiveData
 
+    @RequiresApi(Build.VERSION_CODES.O)
     fun fetchJobs() {
-//        clearJobsList() // xóa item cũ đi trước khi fetch lại
-        FirebaseDatabase.getInstance().getReference("Job")
-            .get().addOnSuccessListener { dataSnapshot->
-                @RequiresApi(Build.VERSION_CODES.O)
-                    for (userSnapshot in dataSnapshot.children) {
+        val uid = GetData.getCurrentUserId()
+        if(uid!= null) {
+            //lấy danh sách Việc đã trúng tuyển
+            approvedJobDb.child(uid).get().addOnSuccessListener { approvedJobSnapshot->
+                val approvedJobIdList: MutableList<String> = mutableListOf()
+                approvedJobSnapshot.children.forEach{
+                    val approvedJobId = it.key.toString()
+                    approvedJobIdList.add(approvedJobId)
+                }
+                database.get().addOnSuccessListener { dataSnapshot ->
+                    dataSnapshot.children.forEach {userSnapshot->
                         val buserId = userSnapshot.key.toString()
                         val tempList: MutableList<JobModel> = mutableListOf()
-                        GetData.getUsernameFromUserId(buserId) { username ->
-                            for (jobSnapshot in userSnapshot.children) {
-                                val jobModel = jobSnapshot.getValue(JobModel::class.java)
-                                jobModel?.let {
-                                    it.BUserName = username.toString()
-                                    it.status = GetData.setStatus(it.startTime.toString(), it.endTime.toString(), it.empAmount.toString(), it.numOfRecruited.toString())
+                        userSnapshot.children.forEach { jobSnapshot ->
+                            val jobModel = jobSnapshot.getValue(JobModel::class.java)
+                            jobModel?.let {
+                                it.status = GetData.setStatus(
+                                    it.startTime.toString(),
+                                    it.endTime.toString(),
+                                    it.empAmount.toString(),
+                                    it.numOfRecruited.toString()
+                                )
 
-                                    if(it.status == "closed"){
-                                        deleteAppliedJob(it.jobId.toString())
-                                    }
-
-                                    tempList.add(it) //Chứa full data toàn bộ các job
-
-                                    if (it.status == "recruiting") { // check trạng thái công việc cho vào viewmodel để hiển thị
-                                        addJobsToJobsList(it)
-                                    }
+                                if (it.status == "closed") {
+                                    deleteAppliedJob(it.jobId.toString())
+                                }
+                                tempList.add(it)
+                                // kiểm tra nếu công việc có trạng thái đang tuyển và không nằm trong danh sách đã trúng tuyển
+                                if (it.status == "recruiting" && !approvedJobIdList.contains(it.jobId.toString())) { // check trạng thái công việc cho vào viewmodel để hiển thị
+                                    addJobsToJobsList(it)
                                 }
                             }
-                            updateStatusToFirebase(buserId,tempList)
-
                         }
+//                        _jobsListLiveData.value = tempList
+                        sortFilter(0, 0, 1, 0.0f, 50000.0f, 0, 24)
+
+                        updateStatusToFirebase(buserId, tempList)
+
                     }
+                }
             }
+        }
     }
 
 
@@ -139,7 +152,7 @@ class FindNewJobViewModel : ViewModel() {
     fun updateStatusToFirebase(userId :String,jobList: List<JobModel>) {
         val updatesMap = mutableMapOf<String, Any?>()
         for (jobModel in jobList) {
-            updatesMap["/${jobModel.jobId}/buserName"] = jobModel.BUserName
+//            updatesMap["/${jobModel.jobId}/buserName"] = jobModel.BUserName
             updatesMap["/${jobModel.jobId}/status"] = jobModel.status
         }
         database.child(userId).updateChildren(updatesMap)
